Integrate change 410140 from main bug: 77958
Add new LC key conversation_ignore_maillist_prefix. When set to true, any prefix inside braces ([]) is ignored when finding matching conversations. When set to false, the prefix in braces is included in the search for related conversations. The default value is true, which keeps the existing behavior intact.

http://bugzilla.zimbra.com/show_bug.cgi?id=77958

Copied from Perforce
 Change: 410141