Bug: 101736
Update timezones.ics to be based on tzdata2015g
Release 2015g changes:
  Turkey's 2015 fall-back transition is scheduled for Nov. 8, not Oct. 25.
      Needed changes described below to reflect this change to Europe/Istanbul
  Norfolk moves from +1130 to +1100 on 2015-10-04 at 02:00 local time.
  Fiji's 2016 fall-back transition is scheduled for January 17, not 24.
      Curiously, this change was already reflected in the previous timezones.ics
  New zone America/Fort_Nelson

Improved handling of upcoming timezone changes. Used to just consider the ZoneLine applicable at the reference date but can get more useful information by also considering any applicable ZoneLine beyond the reference date.
It is quite rare for there to be more than one zoneline (there were only 3 timezones for today's date as the reference date with more than one ZoneLine affecting the future - Europe/Istanbul (3 ZoneLines - but only the first 2 are useful for the simplified zones we produce), America/Grand_Turk and America/Cayman (no changes made for this timezones.ics - see below for reasons)

Europe/Istanbul is currently observing DAYLIGHT time at the same TZOFFSETTO under both the current rules and the next set of rules, so if we just adopt the next set of rules, offsets will be correct under both sets of rules, so that is safe to do.
America/Grand_Turk are adopting STANDARD time at the end of the current DAYLIGHT time but at the same TZOFFSETTO, so it is safe to adopt the next set of rules.
America/Cayman is introducing daylight saving time from next year but to use the new rules now would cause times to be incorrect until the real introduction of the new rules, so deferring that.  If the rules are regenerated for a reference time beyond the first Sunday in November, the new rules will be in place.
https://reviewboard.eng.zimbra.com/r/6409/
JUDASPRIEST ---> main

https://bugzilla.zimbra.com/show_bug.cgi?id=101736

Copied from Perforce
 Change: 557752