Bug: 103295
Added a new jetty filter RestrictedPortFilter in order to restrict access to port 7073 to only those hosts listed in zimbraMailTrustedIP. This filter is situated early on in the filter chain, right after the DoSFilter. The filter has one initialization parameter called "restricted.ports", which is a comma-separated list of ports on which SOAP requests will have their client IPs verified. The only value there right now is that of zimbraMtaAuthPort. If a request comes in on a restricted port from an IP not listed in zimbraMailTrustedIP, a 403 error is thrown and the incident logged in audit.log.

RestrictedPortFilter.java:
 - new filter

web.xml, web.xml.production:
 - defined new filter

jetty-ant-config.xml:
 - added mtaAdminHttpsConnector to jetty debug configuration

https://bugzilla.zimbra.com/show_bug.cgi?id=103295

Copied from Perforce
 Change: 563412